PPS-44: Refactor forms even more
Problem:
* Forms still share a lot of code

Solution:
* Refactor everything into Forms/constants.js

Note:
N/A
