--- src/components/ForgotForm/index.js
@@ -1,12 +1,0 @@
-import React, { Component } from 'react';
-
-import DefaultForm from '../DefaultForm';
-import Forms from '../DefaultForm/constants';
-
-const ForgotForm = () => (
-    <DefaultForm
-        form={Forms.Forgot}
-    />
-);
-
-export default ForgotForm;
