+++ src/components/Forms/stories/LoginForm.stories.js
@@ -1,11 +1,10 @@
 import React from 'react';
 import { withKnobs } from '@storybook/addon-knobs';
 
-import LoginForm from '../index';
-import Popup from '../../Popup';
+import { LoginForm } from '../index';
 
 export default {
-    title: 'Components/LoginForm',
+    title: 'Components/Forms/LoginForm',
     component: LoginForm,
     decorators: [withKnobs],
     argTypes: {
@@ -17,8 +16,3 @@ const Template = (args) => <LoginForm {...args} />;
 
 export const Primary = Template.bind({});
 Primary.args = {};
-
-const TemplateInPopup = (args) => <Popup popupContent={<LoginForm {...args} />} />;
-
-export const InPopup = TemplateInPopup.bind({});
-InPopup.args = {};
