--- src/components/RegisterForm/index.js
@@ -1,12 +1,0 @@
-import React, { Component } from 'react';
-
-import DefaultForm from '../DefaultForm';
-import Forms from '../DefaultForm/constants';
-
-const RegisterForm = () => (
-    <DefaultForm
-        form={Forms.Register}
-    />
-);
-
-export default RegisterForm;
