--- src/components/RegisterForm/stories/RegisterForm.stories.js
@@ -1,24 +1,0 @@
-import React from 'react';
-import { withKnobs } from '@storybook/addon-knobs';
-
-import Popup from '../../Popup';
-import RegisterForm from '../index';
-
-export default {
-    title: 'Components/RegisterForm',
-    component: RegisterForm,
-    decorators: [withKnobs],
-    argTypes: {
-        
-    },
-};
-
-const Template = (args) => <RegisterForm {...args} />;
-
-export const Primary = Template.bind({});
-Primary.args = {};
-
-const TemplateInPopup = (args) => <Popup popupContent={<RegisterForm {...args} />} />;
-
-export const InPopup = TemplateInPopup.bind({});
-InPopup.args = {};
